Create Input and Output Files
In this code, it simulates a scenario where students take random quizzes, and their quiz scores and averages are recorded and analyzed. The results are then stored in output files for further examination. This code uses data generation for the random quiz grades. It also calculates the averages of the quizzes while also taking into account quizzes that have not been attempted. It uses output files for the students information while error checking those files. And lastly, displays the students information with each quiz grade displayed, with the average of them at the end.